Typescriptized
